\name{informRset-package}
\alias{informRset-package}
\alias{informRset}
\docType{package}
\title{
\packageTitle{informRset}
}
\description{
\packageDescription{informRset}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{informRset}
\packageIndices{informRset}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{informRset}

Maintainer: \packageMaintainer{informRset}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
